+++ apps/api/.env.example
@@ -1,3 +1,5 @@
 PORT=4000
 API_CORS_ORIGIN=http://localhost:5173
 NODE_ENV=development
+CLERK_SECRET_KEY=
+CLERK_PUBLISHABLE_KEY=
