+++ apps/api/src/config/env.ts
@@ -10,8 +10,18 @@ const fallback = {
 
 const read = (key: keyof typeof fallback): string => process.env[key] ?? fallback[key];
 
+const readRequired = (key: "CLERK_SECRET_KEY" | "CLERK_PUBLISHABLE_KEY"): string => {
+  const value = process.env[key];
+  if (!value) {
+    throw new Error(`Missing required environment variable: ${key}`);
+  }
+  return value;
+};
+
 export const env = {
   nodeEnv: read("NODE_ENV"),
   port: Number(read("PORT")),
-  corsOrigin: read("API_CORS_ORIGIN")
+  corsOrigin: read("API_CORS_ORIGIN"),
+  clerkSecretKey: readRequired("CLERK_SECRET_KEY"),
+  clerkPublishableKey: readRequired("CLERK_PUBLISHABLE_KEY")
 };
