+++ apps/api/src/routes/index.ts
@@ -1,3 +1,5 @@
+import { ClerkExpressRequireAuth } from "@clerk/express";
+import type { RequestWithAuth } from "@clerk/express";
 import { Router } from "express";
 
 import { healthRouter } from "./health.js";
@@ -6,6 +8,7 @@ export const router = Router();
 
 router.use(healthRouter);
 
-router.get("/v1/ping", (_req, res) => {
-  res.json({ pong: true });
+router.get("/v1/ping", ClerkExpressRequireAuth(), (req: RequestWithAuth, res) => {
+  const { userId } = req.auth;
+  res.json({ pong: true, userId });
 });
