+++ apps/web/.env.example
@@ -1,1 +1,2 @@
 VITE_API_BASE_URL=http://localhost:4000
+VITE_CLERK_PUBLISHABLE_KEY=
