+++ apps/web/src/App.test.tsx
@@ -1,11 +1,11 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
 
-import App from "./App";
+import { MarketingPage } from "./components/MarketingPage";
 
 describe("Landing page", () => {
   it("renders hero content and primary call to action", () => {
-    render(<App />);
+    render(<MarketingPage />);
 
     expect(screen.getByRole("heading", { level: 1, name: /Design, launch, and evolve/i })).toBeInTheDocument();
     expect(screen.getByRole("link", { name: /Start building today/i })).toBeInTheDocument();
@@ -13,7 +13,7 @@ describe("Landing page", () => {
   });
 
   it("includes anchor navigation for key sections", () => {
-    render(<App />);
+    render(<MarketingPage />);
 
     ["Product", "Platform", "Automation", "Contact"].forEach((label) => {
       expect(screen.queryAllByRole("link", { name: label }).length).toBeGreaterThan(0);
