+++ apps/web/src/lib/api.ts
@@ -9,6 +9,11 @@ export type HealthResponse = {
   timestamp: string;
 };
 
+export type PingResponse = {
+  pong: boolean;
+  userId: string;
+};
+
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL ?? "http://localhost:4000";
 
 export const api = {
@@ -19,6 +24,22 @@ export const api = {
         throw new Error(`Request failed with status ${response.status}`);
       }
       const data = (await response.json()) as HealthResponse;
+      return { data };
+    } catch (error) {
+      return { data: null, error: (error as Error).message };
+    }
+  },
+  async ping(token: string): Promise<ApiResult<PingResponse>> {
+    try {
+      const response = await fetch(`${API_BASE_URL}/v1/ping`, {
+        headers: {
+          Authorization: `Bearer ${token}`
+        }
+      });
+      if (!response.ok) {
+        throw new Error(`Request failed with status ${response.status}`);
+      }
+      const data = (await response.json()) as PingResponse;
       return { data };
     } catch (error) {
       return { data: null, error: (error as Error).message };
