+++ apps/web/src/main.tsx
@@ -1,11 +1,21 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
+import { ClerkProvider } from "@clerk/clerk-react";
+
 import App from "./App";
 import "./styles/global.css";
 
+const publishableKey = import.meta.env.VITE_CLERK_PUBLISHABLE_KEY as string | undefined;
+
+if (!publishableKey) {
+  throw new Error("Missing VITE_CLERK_PUBLISHABLE_KEY in environment");
+}
+
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <ClerkProvider publishableKey={publishableKey}>
+      <App />
+    </ClerkProvider>
   </React.StrictMode>
 );
